Big Update, Alls DTO and Services,controller of Yostin

+++ demo/src/main/java/co/edu/udes/activity/backend/demo/models/Subject.java
@@ -11,7 +11,7 @@ public class Subject {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int idSubject;
+    private long idSubject;
 
 
        @Column(name = "name")
